doing some refactoring
