+++ .vscode/launch.json
@@ -9,7 +9,7 @@
             "request": "launch",
             "name": "Launch Program",
             "skipFiles": ["<node_internals>/**"],
-            "program": "${workspaceFolder}\\index.js",
+            "program": "${workspaceFolder}\\src\\index.js",
             "args": ["--experimental-json-modules"],
             "runtimeArgs": ["--experimental-json-modules"]
         }
