+++ src/presence-switcher.js
@@ -1,4 +1,3 @@
-import amongUs from './presences/among-us.js'
 import amongUsPresence from './presences/among-us.js'
 
 export const PresenceSwitcher = (oldPresence, newPresence, store) => {
